times_of_india/requirements.txt
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
jieba3k==0.35.1
joblib==0.15.1
lxml==4.5.1
newspaper3k==0.2.8
nltk==3.5
parsel==1.6.0
Pillow==7.1.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pymongo==3.10.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
PyYAML==5.3.1
queuelib==1.5.0
regex==2020.5.14
requests==2.23.0
requests-file==1.5.1
Scrapy==2.1.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
tinysegmenter==0.3
tldextract==2.2.2
tqdm==4.46.0
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.22.0
zope.interface==5.1.0
